Add clear storage on manage section
