+++ src/components/Home/index.js
@@ -4,7 +4,7 @@ import CrowdsalesList from '../Common/CrowdsalesList'
 import { Loader } from '../Common/Loader'
 import { loadRegistryAddresses } from '../../utils/blockchainHelpers'
 import { ModalContainer } from '../Common/ModalContainer'
-import { toast } from '../../utils/utils'
+import { toast, clearStorage } from '../../utils/utils'
 import { TOAST, NAVIGATION_STEPS, DOWNLOAD_STATUS } from '../../utils/constants'
 import { inject, observer } from 'mobx-react'
 import { checkWeb3, getNetworkVersion } from '../../utils/blockchainHelpers'
@@ -55,15 +55,15 @@ export class Home extends Component {
     try {
       await loadRegistryAddresses()
       this.setState({
-        loading: false,
         showModal: true
       })
     } catch (e) {
       logger.error('There was a problem loading the crowdsale addresses from the registry', e)
-      this.setState({
-        loading: false
-      })
     }
+
+    this.setState({
+      loading: false
+    })
   }
 
   goNextStep = async () => {
@@ -71,7 +71,7 @@ export class Home extends Component {
     if (storage.has('DeploymentStore') && storage.get('DeploymentStore').deploymentStep) {
       this.props.history.push('/')
     } else {
-      this.clearStorage()
+      clearStorage(this.props)
       await this.reloadStorage()
       this.props.history.push('1')
     }
@@ -96,41 +96,6 @@ export class Home extends Component {
           'The contracts could not be downloaded.Please try to refresh the page. If the problem persists, try again later.'
       })
       contractStore.setProperty('downloadStatus', DOWNLOAD_STATUS.FAILURE)
-    }
-  }
-
-  clearStorage() {
-    // Generate of stores to clear
-    const toArray = ({
-      generalStore,
-      contractStore,
-      crowdsaleStore,
-      gasPriceStore,
-      deploymentStore,
-      reservedTokenStore,
-      stepTwoValidationStore,
-      tierStore,
-      tokenStore
-    }) => {
-      return [
-        generalStore,
-        contractStore,
-        crowdsaleStore,
-        gasPriceStore,
-        deploymentStore,
-        reservedTokenStore,
-        stepTwoValidationStore,
-        tierStore,
-        tokenStore
-      ]
-    }
-
-    const storesToClear = toArray(this.props)
-    for (let storeToClear of storesToClear) {
-      if (typeof storeToClear.reset === 'function') {
-        logger.log('Store to be cleared:', storeToClear.constructor.name)
-        storeToClear.reset()
-      }
     }
   }
